Remove vestiges of the old build system, update TOC sidebar styling
