+++ lib/md.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import _ from 'lodash';
 import markdownIt from 'markdown-it';
-import { translateLink } from './utils.js';
+import URI from 'urijs';
 
 export function makeRenderer(toc) {
     const md = markdownIt({ html: true, linkify: true });
@@ -105,4 +105,33 @@ export function extractVersionChanges(ver, contents) {
     });
 
     return output.join("\n");
+}
+
+// Translates links to .md files (usually pointing into the mercurial tree)
+// to links within *this* site, if possible
+export function translateLink(href) {
+    var uri = URI(href);
+    var path = uri.path();
+    var slash = path.lastIndexOf("/");
+    var fileName = path.substring(slash + 1);
+    if (_.endsWith(fileName, '.md')) {
+        if (uri.hostname() && uri.hostname().indexOf('sourceforge') < 0) {
+            return href;
+        }
+
+        uri = uri.hostname(null).protocol(null);
+
+        if (fileName === "README.md") {
+            return uri.path('/index.html').toString();
+        } else {
+            return uri.path(replaceExtension(fileName, '.html')).toString();
+        }
+    }
+
+    return href;
+}
+
+function replaceExtension(filename, newext) {
+    const ext = path.extname(filename);
+    return ext.slice(0, filename.length - ext.length) + newext;
 }
