+++ src/_includes/docs.njk
@@ -5,10 +5,9 @@ layout: "content.njk"
   <div class="col-sm-8" id="md-contents">
     {{ content | stripTOC | safe }}
   </div>
-  <div class="col-sm-4 offset-sm-8 sidebar hidden-xs-down h-100" style="position:fixed; overflow-x:hidden;">
-    <div class="" id="md-toc">
-    </div>
-  </nav>
+  <div class="col-sm-4 d-none d-sm-block">
+    <nav class="pt-3 sticky-top" id="md-toc"></nav>
+  </div>
 </div>
 
 {% js "defer" %}
@@ -20,7 +19,6 @@ tocbot.init({
   tocSelector: '#md-toc',
   contentSelector: '#md-contents',
   headingSelector: 'h2, h3, h4',
-  collapseDepth: 5,
   scrollContainer: 'body',
   scrollSmooth: true,
 });
@@ -31,7 +29,6 @@ layoutbus.addEventListener("navResize", ev => {
     tocSelector: '#md-toc',
     contentSelector: '#md-contents',
     headingSelector: 'h2, h3, h4',
-    collapseDepth: 5,
     scrollContainer: 'body',
     scrollSmooth: true,
     headingsOffset: height + 5,
@@ -39,7 +36,3 @@ layoutbus.addEventListener("navResize", ev => {
   });
 });
 {% endjs %}
-
-{% css "defer" %}
-<link rel="stylesheet" href="/css/tocbot.css">
-{% endcss %}
