+++ src/css/main.scss
@@ -31,6 +31,11 @@ h2, h3, .nav-target {
 @import "tocbot/src/scss/_tocbot-core"; 
 
 #md-toc {
+  top: var(--nav-height, 85px);
+  max-height: calc(100vh - var(--nav-height, 85px));
+  overflow-y: auto;
+  scrollbar-width: none;
+
   // Wipe out browser layout numbers completely
   ol.toc-list {
     list-style-type: none;
@@ -59,5 +64,13 @@ h2, h3, .nav-target {
     .toc-link {
       font-size: $font-size-base * 0.85;
     }
+  }
+
+  // Dynamic Scrollspy Active State Class handled by Tocbot
+  .is-active-link {
+    font-weight: $font-weight-semibold;
+    color: $primary;
+    border-left-color: $primary; // Reveals the sidebar accent indicator
+    background-color: rgba($primary, 0.04);
   }
 }
